+++ CHANGELOG.md
@@ -16,6 +16,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Make possible to start the backend-server without orchestrator connection available (#811)
 - Use the connect-tools image 0.10.0 in the example yaml files (#850)
 
+### Fixed
+- Compute CP dates after updating related tasks (#855)
+
 ## [0.9.0] 2022-03-01
 
 ### Added